tests/wpt/tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/stable-state-dialogs-manual.html

<!doctype html>
<title>stable state in dialogs</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
['alert', 'confirm', 'prompt'].forEach(function(dialog) {
  test(function() {
    v = document.createElement('video');
    v.src = 'data:,';
    assert_equals(v.networkState, v.NETWORK_NO_SOURCE, 'networkState before dialog');
    assert_equals(v.currentSrc, '', 'currentSrc before dialog');
    window[dialog]('dismiss this dialog');
    assert_equals(v.networkState, v.NETWORK_NO_SOURCE, 'networkState after dialog');
    assert_equals(v.currentSrc, '', 'currentSrc after dialog');
  }, 'stable state in ' + dialog + '()');
});
</script>
