third_party/WebKit/LayoutTests/imported/wpt/html/semantics/embedded-content/media-elements/loading-the-media-resource/stable-state-dialogs-manual-expected.txt

ALERT: dismiss this dialog
CONFIRM: dismiss this dialog
PROMPT: dismiss this dialog, default text: 
This is a testharness.js-based test.
PASS stable state in alert() 
PASS stable state in confirm() 
PASS stable state in prompt() 
Harness: the test ran to completion.
